backend/requirements-agentic.txt
# ============================================================================
# Pavilion-AI Agentic CMS - Python Dependencies
# ============================================================================
# Install with: pip install -r requirements-agentic.txt
#
# This file includes all dependencies for the Agentic CMS implementation
# with Vertex AI and Google Cloud Platform integration
# ============================================================================

# ============================================================================
# Core Framework & Web
# ============================================================================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
django-environ==0.21.0
python-decouple==3.8
drf-spectacular==0.26.5  # API documentation

# ============================================================================
# Database & ORM
# ============================================================================
psycopg2-binary==2.9.9  # PostgreSQL adapter
psycopg==3.1.14  # PostgreSQL async
SQLAlchemy==2.0.23
alembic==1.13.1

# ============================================================================
# Google Cloud Platform
# ============================================================================
google-cloud-aiplatform==1.35.0
google-cloud-storage==2.14.0
google-cloud-sql-connector==1.4.6
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0
google-cloud-trace==1.11.1
google-cloud-tasks==2.14.1
google-cloud-texttospeech==2.14.2
google-cloud-vision==3.6.0
google-api-python-client==2.109.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# ============================================================================
# AI/ML Libraries
# ============================================================================
langchain==0.1.14
langchain-google-vertexai==0.1.3
langchain-community==0.0.32
langchain-text-splitters==0.0.1
google-generativeai==0.3.5
vertexai==0.1.1  # Vertex AI SDK
anthropic==0.7.1  # If using Claude API
openai==1.3.10  # OpenAI (for compatibility/migration)

# ============================================================================
# LLM & Vector Database
# ============================================================================
sentence-transformers==2.2.2  # For local embeddings (backup)
chromadb==0.4.24  # Local vector store (development)
pinecone-client==3.0.2  # If migrating to Pinecone later
qdrant-client==2.7.0  # If migrating to Qdrant later

# ============================================================================
# Task Queue & Background Jobs
# ============================================================================
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# ============================================================================
# Content & Text Processing
# ============================================================================
beautifulsoup4==4.12.2
lxml==4.9.3
markdown==3.5.1
bleach==6.1.0
pillow==10.1.0  # Image processing
PyYAML==6.0.1
python-slugify==8.0.1

# ============================================================================
# Authentication & Security
# ============================================================================
PyJWT==2.8.1
djangorestframework-simplejwt==5.3.2
cryptography==41.0.7
django-allauth==0.57.0

# ============================================================================
# Utilities & Helpers
# ============================================================================
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
tenacity==8.2.3  # Retry library
pydantic==2.5.2  # Data validation
python-dateutil==2.8.2
pytz==2023.3.post1
attrs==23.2.0

# ============================================================================
# Monitoring & Logging
# ============================================================================
structlog==23.3.0
python-json-logger==2.0.7
sentry-sdk==1.39.1

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.23.2
factory-boy==3.3.0
faker==22.0.0
coverage==7.4.0
responses==0.24.1

# ============================================================================
# Development Tools
# ============================================================================
black==23.12.1  # Code formatter
flake8==6.1.0  # Linter
isort==5.13.2  # Import sorter
mypy==1.7.1  # Type checker
django-extensions==3.2.3
django-debug-toolbar==4.2.0
ipython==8.18.1
ipdb==0.13.13

# ============================================================================
# Documentation
# ============================================================================
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# ============================================================================
# Admin & Management
# ============================================================================
django-admin-interface==0.26.0

# ============================================================================
# Optional: Media Processing (Phase 5)
# ============================================================================
# Uncomment if implementing video/image generation
# moviepy==1.0.3
# ffmpeg-python==0.2.1
# opencv-python==4.8.1.78

# ============================================================================
# Production Server
# ============================================================================
gunicorn==21.2.0
whitenoise==6.6.0  # Static file serving
uvicorn==0.24.0  # ASGI server (for async)

# ============================================================================
# Database Utilities
# ============================================================================
django-db-geventpool==4.0.1  # Connection pooling
django-cachalot==2.6.1  # ORM caching

# ============================================================================
# Notes
# ============================================================================
# Install with: pip install -r backend/requirements-agentic.txt
#
# For development, also install:
#   pip install -r backend/requirements-dev.txt
#
# Create a specific version lock file:
#   pip freeze > requirements.lock
#
# Key Vertex AI versions:
# - google-cloud-aiplatform: 1.35.0+
# - google-generativeai: 0.3.5+ (for latest models)
# - langchain-google-vertexai: 0.1.3+
#
# Async support:
# - For async Vertex AI: use vertexai 0.1.1+
# - For async database: use psycopg (not psycopg2)
